feat: 使用子类来辅助核心代码开发
