+++ vnpy/event/event.py
@@ -1,0 +1,7 @@
+# encoding: UTF-8
+
+from vnpy.event.oevent import Event as vtEvent
+
+class Event(vtEvent):
+    pass
+
