+++ vnpy/event/oevent.py
@@ -1,0 +1,10 @@
+# encoding: UTF-8
+
+class Event:
+    """事件对象"""
+
+    #----------------------------------------------------------------------
+    def __init__(self, type_=None):
+        """Constructor"""
+        self.type_ = type_      # 事件类型
+        self.dict_ = {}         # 字典用于保存具体的事件数据
