+++ vnpy/event/seventEngine.py
@@ -1,0 +1,9 @@
+# encoding: UTF-8
+
+from vnpy.event.oeventEngine import EventEngine2
+from vnpy.event.oeventEngine import Event as vtEvent
+
+class EventEngine(EventEngine2):
+    """
+    在这个子类中进行核心代码更改
+    """
